Started using LuaAutoC for bindings
Achieved mutable C structs that are accessible from Lua. However, struct
nesting is still out of reach as it will take more time to build with
acceptable quality than the time available for this assignment.

N.B. LuaAutoC might not be the best choice for evol as it has
performance overhead everywhere. While the overhead is well-justified
and is what helps achieve the flexibility of LuaAutoC, this is probably
a trade-off that should not be accepted in evol. More research is
needed.

Signed-off-by: Robear Selwans <robear.selwans@outlook.com>